+++ lib/gg.dart
@@ -29,6 +29,7 @@ export 'src/commands/do/do_publish.dart';
 export 'src/commands/did.dart';
 export 'src/commands/did/did_push.dart';
 export 'src/commands/did/did_commit.dart';
+export 'src/commands/did/did_publish.dart';
 
 // Tools
 export 'src/tools/carriage_return.dart';